feat: moved report_pdf under /src/strategy_tester/report/pdf folder
