+++ src/strategy_tester/metrics/__init__.py
@@ -28,5 +28,5 @@ from .profits_losses_bars import get_profits_losses_by_hour, get_profits_losses_
 from .profits_losses_by_bar_opened import get_profits_by_time_opened, get_losses_by_time_opened
 from .profits_losses_mean import get_profits_losses_mean_by_hour, get_profits_losses_mean_by_dayofweek, get_profits_losses_mean_by_month
 from .profits_losses_sum import get_profits_losses_sum_by_hour, get_profits_losses_sum_by_dayofweek, get_profits_losses_sum_by_month
-from .report_pdf import plot_metrics, report_to_pdf
+from ..report.pdf.report_pdf import plot_metrics, report_to_pdf
 from .save import save_metrics
