+++ src/strategy_tester/report/__init__.py
@@ -1,0 +1,6 @@
+"""
+Module with helper functions to make report files.
+
+Inside this module you can find:\n
++ `report_to_pdf` to save statistics, metrics and so on as pdf, see `.pdf.report_to_pdf`\n
+"""
